Move readability stat computation to quanteda
Previously, this was done locally by a function that duplicated much of the machinery of `quanteda::textstat_readability()`.  `covars_make()` now uses that function in **quanteda** for computing the intermediate quantities.  This includes the fix to the computation of the word list-based measures such as the Dale-Chall unfamiliar words.  See https://github.com/quanteda/quanteda/pull/1425.